Add useEditorOperations hook
